requirements.txt
# 财务数据分析工具 - 依赖包
pandas>=1.5.0
openpyxl>=3.0.0
PyQt5>=5.15.0
PyYAML>=6.0
pyinstaller>=5.0
